Extract options into own file
