+++ src/index.ts
@@ -2,6 +2,7 @@ import { AST, Doc, FastPath, format, Options, Parser, ParserOptions, Plugin, uti
 // @ts-ignore
 import * as lex from 'pug-lexer';
 import { createLogger, Logger, LogLevel } from './logger';
+import { options as pugOptions, PugParserOptions, resolveAttributeSeparatorOption } from './options';
 import { AttributeToken, EndAttributesToken, Token } from './pug-token';
 
 const { makeString } = util;
@@ -9,24 +10,6 @@ const { makeString } = util;
 const logger: Logger = createLogger(console);
 if (process.env.NODE_ENV === 'test') {
 	logger.setLogLevel(LogLevel.DEBUG);
-}
-
-const CATEGORY_PUG: string = 'Pug';
-
-export interface PugParserOptions {
-	attributeSeparator: 'always' | 'as-needed';
-}
-
-function resolveAttributeSeparatorOption(attributeSeparator: 'always' | 'as-needed'): boolean {
-	switch (attributeSeparator) {
-		case 'always':
-			return true;
-		case 'as-needed':
-			return false;
-	}
-	throw new Error(
-		`Invalid option for pug attributeSeparator. Found '${attributeSeparator}'. Possible options: 'always' or 'as-needed'`
-	);
 }
 
 type QuotationType = 'SINGLE' | 'DOUBLE';
@@ -585,27 +568,7 @@ export const plugin: Plugin = {
 			}
 		}
 	},
-	options: {
-		attributeSeparator: {
-			since: '1.0.0',
-			category: CATEGORY_PUG,
-			type: 'choice',
-			default: 'always',
-			description: 'Change when attributes are separated by commas in tags.',
-			choices: [
-				{
-					value: 'always',
-					description:
-						'Always separate attributes with commas. Example: `button(type="submit", (click)="play()", disabled)`'
-				},
-				{
-					value: 'as-needed',
-					description:
-						'Only add commas between attributes where required. Example: `button(type="submit", (click)="play()" disabled)`'
-				}
-			]
-		}
-	} as any,
+	options: pugOptions as any,
 	defaultOptions: {}
 };
